Only delete secret if env has value

+++ packages/now-cli/src/util/env/remove-env-record.ts
@@ -22,17 +22,17 @@ export default async function removeEnvRecord(
     method: 'DELETE',
   });
 
-  const idOrName = (env.value || '').startsWith('@')
-    ? env.value.slice(1)
-    : env.value;
-  const urlSecret = `/v2/now/secrets/${idOrName}`;
-  const secret = await client.fetch<Secret>(urlSecret);
+  if (env && env.value) {
+    const idOrName = env.value.startsWith('@') ? env.value.slice(1) : env.value;
+    const urlSecret = `/v2/now/secrets/${idOrName}`;
+    const secret = await client.fetch<Secret>(urlSecret);
 
-  // Since integrations add global secrets, we must only delete if the secret was
-  // specifically added to this project
-  if (secret && secret.projectId === projectId) {
-    await client.fetch<Secret>(urlSecret, {
-      method: 'DELETE',
-    });
+    // Since integrations add global secrets, we must only delete if the secret was
+    // specifically added to this project
+    if (secret && secret.projectId === projectId) {
+      await client.fetch<Secret>(urlSecret, {
+        method: 'DELETE',
+      });
+    }
   }
 }
